refact : activity to fragment and framelayout
