+++ pooni/app/src/main/java/com/uki121/pooni/FragmentHomeMenu.java
@@ -1,6 +1,7 @@
 package com.uki121.pooni;
 
 import android.app.Activity;
+import android.app.DialogFragment;
 import android.app.Fragment;
 import android.app.FragmentTransaction;
 import android.content.Context;
@@ -20,7 +21,7 @@ import com.afollestad.materialdialogs.MaterialDialog;
 
 public class FragmentHomeMenu extends Fragment {
 
-    private View setting_dial_view;
+    //private View setting_dial_view;
     private bookShelf sb;
 
 
@@ -28,8 +29,8 @@ public class FragmentHomeMenu extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
+
         final View view = inflater.inflate(R.layout.fragment_home, container, false);
-        sb = new bookShelf();
 
         Button btn_start = (Button) view.findViewById(R.id.btn_new_start);
         Button btn_quick = (Button) view.findViewById(R.id.btn_quick_start);
@@ -39,33 +40,7 @@ public class FragmentHomeMenu extends Fragment {
             @Override
             public void onClick(View v) {
                 // Do something in response to button click
-                boolean wrapInScrollView = true;
-                Fragment newFragment = new FragmentHomeMenu();
-                final MaterialDialog startDialog = new MaterialDialog.Builder(newFragment.getActivity())
-                        .title("setting")
-                        .positiveText("확인")
-                        .negativeText("취소")
-                        .onPositive(new MaterialDialog.SingleButtonCallback() {
-                            @Override
-                            public void onClick(@NonNull MaterialDialog dialog, @NonNull DialogAction which) {
-                                // Some unimportant data stuff
-                                /* Todo_BIG*/
-                                try {
-                                    setting_dial_view = dialog.getCustomView();
-                                    if (setting_dial_view != null)
-                                    {
-                                        //View v = dialog.getCustomView();
-                                        sb.AddBooks(setting_dial_view);
-                                        sb.printBooks();
-                                    }
-                                } catch (NullPointerException e) {
-                                    System.out.println(e.getMessage());
-                                }
-                            }
-                        })
-                        .customView(R.layout.customview_setting, wrapInScrollView)
-                        .build();
-                startDialog.show();
+                openDialog();
             }
         });
         btn_quick.setOnClickListener(new View.OnClickListener() {
@@ -100,5 +75,10 @@ public class FragmentHomeMenu extends Fragment {
         });
 
         return view;
+    }
+    private void openDialog() {
+        DialogFragment setDialogFragment = new dialogCustomSet();
+        setDialogFragment.setTargetFragment(this, 0);
+        setDialogFragment.show(getFragmentManager(), "SweetAlert");
     }
 }
